Fix NullPointerException with Multiverse-Core
Added "softdepend : [Multiverse-Core]" to the plugin.yml in order to avoid that the plugin generates a NullPointerException because Multiverse-Core has not loaded the worlds yet. The general role of the "softdepend" parameter is to ask the server to load the plugin after a certain list of plugins (if they are present).